patch for new fixes i made before
remade system of reading the formant settings on startup (still janky as hell, going to remake with sql) fixed when value is true in the txt but the checkbox isn't checked. therefore, also fixed when other parts of formanting in the ui doesn't appear.

my_utils.py now utilizes the stftpitchshift from local environment instead of globally installed one.