fix
Add a check function to filter out devices that are not working.
Add a function to remove parametrizations from the model.
Disable clipping warning.
Add a requirement to agree to the terms in order to use real-time mode.
Replace the indexing method of the audio devices.
Fix the issue where the label does not respond to messages.